+++ src/nushu/set-comments.js
@@ -3,6 +3,7 @@ const fs = require("node:fs/promises");
 const lightColors = require("@primer/primitives/dist/json/colors/light.json");
 const darkColors = require("@primer/primitives/dist/json/colors/dark.json");
 const { flagHex } = require("./util");
+const chroma = require("chroma-js");
 
 /*
 Set comments to green 4 & 6, 90% alpha
@@ -11,7 +12,7 @@ Set comments to green 4 & 6, 90% alpha
 async function setLightComments() {
   const lightTheme = JSON.parse(await fs.readFile("./themes/nushu-light.json"));
   const tokenColors = lightTheme.tokenColors;
-  const green = lightColors.scale.green[5] + "E6"; // 90%
+  const green = chroma(lightColors.scale.green[5]).alpha(0.9).hex(); // 90%
 
   const [commentToken] = tokenColors.filter(
     (token) => token.scope.indexOf("comment") >= 0
@@ -32,7 +33,7 @@ async function setLightComments() {
 async function setDarkComments() {
   const darkTheme = JSON.parse(await fs.readFile("./themes/nushu-dark.json"));
   const tokenColors = darkTheme.tokenColors;
-  const green = darkColors.scale.green[3] + "E6"; // 80%
+  const green = chroma(darkColors.scale.green[3]).alpha(0.9).hex(); // 90%
 
   const [commentToken] = tokenColors.filter(
     (token) => token.scope.indexOf("comment") >= 0