+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.1.20
+
+- Nüshu Light/Dark: Use slightly transparent text for begin/end text strings, so the content that matters pops a bit
+- Nüshu Light/Dark: Set unset specific borders explicitly
+
 ## 2.1.19
 
 - Nüshu Light/Dark: Use slightly transparent borders in standard cases